requirements.txt
torch
torchvision
tensorboard
pyyaml
tqdm
numpy
torchmetrics
gdown